+++ src/RcppExports.cpp
@@ -8,7 +8,7 @@ using namespace Rcpp;
 
 // qloop_numeric
 NumericVector qloop_numeric(NumericVector times, NumericVector service, int n_servers);
-RcppExport SEXP queuecomputer_qloop_numeric(SEXP timesSEXP, SEXP serviceSEXP, SEXP n_serversSEXP) {
+RcppExport SEXP _queuecomputer_qloop_numeric(SEXP timesSEXP, SEXP serviceSEXP, SEXP n_serversSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -21,7 +21,7 @@ END_RCPP
 }
 // qloop_qq
 NumericVector qloop_qq(NumericVector times, NumericVector service, NumericVector x, NumericVector y);
-RcppExport SEXP queuecomputer_qloop_qq(SEXP timesSEXP, SEXP serviceSEXP, SEXP xSEXP, SEXP ySEXP) {
+RcppExport SEXP _queuecomputer_qloop_qq(SEXP timesSEXP, SEXP serviceSEXP, SEXP xSEXP, SEXP ySEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;